pyproject.toml
[tool.poetry]
name = "projecta"
version = "0.1.0"
description = ""
authors = ["Javier de Santiago <jasampa3000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
result = "^0.15.0"
requests = "^2.31.0"
pygithub = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
